requirements.txt
# Dépendances de base
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.5.3
aiohttp==3.9.3
jinja2==3.1.3
python-dotenv==1.0.0
websockets>=12.0.0  # Pour les communications en temps réel
httpx>=0.24.0  # Client HTTP asynchrone pour les appels API

# LLM et Embeddings
anthropic==0.18.0
voyageai==0.3.2

# Traitement PDF et texte
PyMuPDF==1.23.8  # Meilleure performance pour les gros PDFs
python-magic>=0.4.24,<0.5.0
nltk>=3.6.0,<4.0.0
langdetect>=1.0.9  # Pour la détection de langue des documents

# OCR et traitement d'images
ocrmypdf>=15.0.0  # Pour l'OCR des PDFs scannés (nécessite Tesseract, Poppler et Ghostscript)
pdf2image>=1.16.0  # Pour convertir les PDFs en images
Pillow>=10.0.0  # Pour le traitement d'images et les diagnostics OCR

# Base de données vectorielle
qdrant-client==1.7.3

# Monitoring et performance
psutil==5.9.8
tenacity==8.2.3  # Pour les mécanismes de retry avec backoff
rich==13.7.0  # Pour l'affichage amélioré des diagnostics
tqdm>=4.66.0  # Pour les barres de progression lors des longs traitements

# Traitement par lots et parallélisation (pour résoudre les timeouts)
celery>=5.3.0  # Pour le traitement asynchrone des tâches longues
redis>=5.0.0   # Comme broker/backend pour Celery
more-itertools>=10.0.0  # Utilities pour le traitement par lots
backoff>=2.2.1  # Pour implémenter des stratégies de retry avancées

# Tests et développement
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black>=22.0.0,<23.0.0
flake8>=4.0.0,<5.0.0
